Move Kdump client to lib
To avoid executing it at the time to require the file. To do it less
risky, the namespace has not been changed by now.
